pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "windsurf-agent"
version = "0.1.0"
description = "A RAG framework for Windsurf AI models with Cloudera integration"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Ian Brooks", email = "ibrooks@cloudera.com" },
]
dependencies = [
    "requests>=2.25.0",
    "numpy>=1.19.0",
    "pydantic>=1.8.0",
    "python-dotenv>=0.19.0",
    "tqdm>=4.65.0",
    "tenacity>=8.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "types-requests>=2.28.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=windsurf_agent --cov-report=term-missing"